@@ -1,11 +1,10 @@
-import { SQS } from "aws-sdk";
-import { TopicsList, Subscription } from "aws-sdk/clients/sns";
+import { TopicsList, Subscription } from "aws-sdk/clients/sns.js";
 import fetch from "node-fetch";
 import { URL } from "url";
-import { IDebug, ISNSServer } from "./types";
-import * as bodyParser from "body-parser";
-import * as _ from "lodash";
-import * as xml from "xml";
+import { IDebug, ISNSServer } from "./types.js";
+import bodyParser from "body-parser";
+import _ from "lodash";
+import xml from "xml";
 import {
   arrayify,
   createAttr,
@@ -17,7 +16,8 @@ import {
   validatePhoneNumber,
   topicArnFromName,
   formatMessageAttributes,
-} from "./helpers";
+} from "./helpers.js";
+import { SQSClient, SendMessageCommand } from "@aws-sdk/client-sqs";
 
 export class SNSServer implements ISNSServer {
   private topics: TopicsList;
@@ -92,7 +92,8 @@ export class SNSServer implements ISNSServer {
               req.body.Subject,
               req.body.Message,
               req.body.MessageStructure,
-              parseMessageAttributes(req.body)
+              parseMessageAttributes(req.body),
+              req.body.MessageGroupId
             )
           )
         );
@@ -256,9 +257,9 @@ export class SNSServer implements ISNSServer {
       if (_.intersection(v as unknown[], attrs).length > 0) {
         this.debug(
           "filterPolicy Passed: " +
-            v +
-            " matched message attrs: " +
-            JSON.stringify(attrs)
+          v +
+          " matched message attrs: " +
+          JSON.stringify(attrs)
         );
         shouldSend = true;
       } else {
@@ -269,9 +270,9 @@ export class SNSServer implements ISNSServer {
     if (!shouldSend) {
       this.debug(
         "filterPolicy Failed: " +
-          JSON.stringify(policies) +
-          " did not match message attrs: " +
-          JSON.stringify(messageAttrs)
+        JSON.stringify(policies) +
+        " did not match message attrs: " +
+        JSON.stringify(messageAttrs)
       );
     }
 
@@ -282,42 +283,53 @@ export class SNSServer implements ISNSServer {
     return fetch(sub.Endpoint, {
       method: "POST",
       body: event,
-      timeout: 0,
       headers: {
         "x-amz-sns-rawdelivery": "" + raw,
         "Content-Type": "text/plain; charset=UTF-8",
-        "Content-Length": Buffer.byteLength(event),
+        "Content-Length": Buffer.byteLength(event).toString(),
       },
     })
       .then((res) => this.debug(res))
       .catch((ex) => this.debug(ex));
   }
 
-  private publishSqs(event, sub, messageAttributes) {
+  private publishSqs(event, sub, messageAttributes, messageGroupId) {
     const subEndpointUrl = new URL(sub.Endpoint);
     const sqsEndpoint = `${subEndpointUrl.protocol}//${subEndpointUrl.host}/`;
-    const sqs = new SQS({ endpoint: sqsEndpoint, region: this.region });
+    const sqs = new SQSClient({ endpoint: sqsEndpoint, region: this.region });
 
     if (sub["Attributes"]["RawMessageDelivery"] === "true") {
-      return sqs
-        .sendMessage({
-          QueueUrl: sub.Endpoint,
-          MessageBody: event,
-          MessageAttributes: formatMessageAttributes(messageAttributes),
-        })
-        .promise();
+      const sendMsgReq = new SendMessageCommand({
+        QueueUrl: sub.Endpoint,
+        MessageBody: event,
+        MessageAttributes: formatMessageAttributes(messageAttributes),
+        ...(messageGroupId && { MessageGroupId: messageGroupId }),
+      });
+      return new Promise<void>((resolve, reject) => {
+        sqs
+          .send(sendMsgReq).then(() => {
+            resolve();
+          });
+      });
     } else {
       const records = JSON.parse(event).Records ?? [];
       const messagePromises = records.map((record) => {
-        return sqs
-          .sendMessage({
-            QueueUrl: sub.Endpoint,
-            MessageBody: JSON.stringify(record.Sns),
-            MessageAttributes: formatMessageAttributes(messageAttributes),
-          })
-          .promise();
+        const sendMsgReq = new SendMessageCommand({
+          QueueUrl: sub.Endpoint,
+          MessageBody: JSON.stringify(record.Sns),
+          MessageAttributes: formatMessageAttributes(messageAttributes),
+          ...(messageGroupId && { MessageGroupId: messageGroupId }),
+        });
+        return new Promise<void>((resolve, reject) => {
+          sqs
+            .send(sendMsgReq).then(() => {
+              resolve();
+            });
+        });
       });
-      return Promise.all(messagePromises);
+      return new Promise<void>((resolve, reject) => {
+        Promise.all(messagePromises).then(() => resolve());
+      });
     }
   }
 
@@ -326,7 +338,8 @@ export class SNSServer implements ISNSServer {
     subject,
     message,
     messageStructure,
-    messageAttributes
+    messageAttributes,
+    messageGroupId
   ) {
     const messageId = createMessageId();
     Promise.all(
@@ -355,7 +368,8 @@ export class SNSServer implements ISNSServer {
                 subject,
                 message,
                 messageId,
-                messageAttributes
+                messageAttributes,
+                messageGroupId
               )
             );
           }
@@ -368,7 +382,12 @@ export class SNSServer implements ISNSServer {
             return this.publishHttp(event, sub, isRaw);
           }
           if (protocol === "sqs") {
-            return this.publishSqs(event, sub, messageAttributes);
+            return this.publishSqs(
+              event,
+              sub,
+              messageAttributes,
+              messageGroupId
+            );
           }
           throw new Error(
             `Protocol '${protocol}' is not supported by serverless-offline-sns`
@@ -411,7 +430,7 @@ export class SNSServer implements ISNSServer {
     if (msg instanceof Object) {
       try {
         msg = JSON.stringify(msg);
-      } catch (ex) {}
+      } catch (ex) { }
     }
     this.pluginDebug(msg, "server");
   }
